pyproject.toml
[project]
name = "coral-ml-mentor-agent"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
"langchain==0.3.25",
"langchain-community==0.3.24",
"langchain-experimental==0.3.4",
"langchain-mcp-adapters==0.1.7",
"langchain-openai==0.3.26",
"python-dotenv>=1.0.0",
"uv>=0.7.17",
]